requirements.txt
requests
tqdm
readability-lxml
lxml_html_clean
